requirements.txt
django-admin-interface
django-colorfield
django-versatileimagefield
django-money
django-tinymce
django-import-export
django-crispy-forms
crispy-bootstrap5
django-tables2
django-filter
django-registration-redux
django-extensions
django-rosetta
python-decouple
django-user-sessions
django==4.2
psycopg2-binary
pdfkit
pytz